+++ src/array/guard.ts
@@ -1,0 +1,3 @@
+import { createSafeOperation } from '../safe';
+
+export const safeArr = createSafeOperation<Array<any>>([]);
