+++ src/array/index.ts
@@ -1,0 +1,12 @@
+export { concat, includes } from '../generic';
+
+export * from './actions';
+export * from './every';
+export * from './filter';
+export * from './guard';
+export * from './join';
+export * from './map';
+export * from './reduce';
+export * from './slice';
+export * from './some';
+export * from './sort';
